refactor: finish github integration and move google and github to own files
