+++ lib/session.js
@@ -1,6 +1,5 @@
 "use strict";
 
-let crypto = require("crypto");
 let bodyParser = require("body-parser");
 let cookieParser = require("cookie-parser");
 
@@ -64,86 +63,19 @@ if (keypair.private) {
 }
 */
 
-function verifyGitHubToken(/*verifyOpts*/) {
-  return async function (req, res, next) {
-    let request = require("@root/request");
-    let token = (req.headers.authorization || "").replace(/^Bearer /, "");
-
-    // See https://docs.github.com/en/rest/reference/users
-    // notably: name (given + family? arbitrary?), email,
-    // login (suggested username), avatar_url
-
-    // SECURITY: You MUST manually check if an email address is verified:
-    // https://docs.github.com/en/rest/reference/users#emails
-    let resp2 = await request({
-      //url: "https://api.github.com/user",
-      url: "https://api.github.com/user/emails",
-      headers: {
-        Accept: "application/vnd.github.v3+json",
-        Authorization: "Token " + token,
-      },
-      json: true,
-    });
-
-    /** @typedef GhEmail
-     * @property {string} email
-     * @property {boolean} primary
-     * @property {boolean} verified
-     */
-
-    /** @type {Array<GhEmail>} */
-    let emails = resp2.toJSON().body;
-    let email;
-    if (Array.isArray(emails)) {
-      email = emails
-        ?.filter(function (identifier) {
-          return identifier.verified && identifier.primary;
-        })
-        .map(function (identifier) {
-          return identifier.email;
-        })[0];
-    }
-    if (!email) {
-      throw E.UNVERIFIED_OIDC_IDENTIFIER("email");
-    }
-
-    req._oauth2 = {
-      //id: id,
-      //name: name,
-      email: email,
-      iss: "https://github.com",
-    };
-    next();
-  };
-}
-
-function verifyGoogleToken(clientId, verifyOpts) {
-  if (!verifyOpts) {
-    verifyOpts = {};
-  }
-  verifyOpts.iss = "https://accounts.google.com";
-  return verifyOidcToken(verifyOpts, async function verifier(jws) {
-    if (jws.claims.azp != clientId) {
-      throw E.SUSPICIOUS_TOKEN();
-    }
-    if (!jws.claims.email_verified) {
-      throw E.UNVERIFIED_OIDC_IDENTIFIER("email");
-    }
-  });
-}
-
 function verifyOidcToken(verifyOpts, verifier) {
   if (!verifier) {
     verifier = async function () {};
   }
-  // Only tokens signed by accounts.google.com are valid
 
+  // Only tokens signed by the expected issuer (such as https://accounts.google.com) are valid
   return function (req, res, next) {
     let jwt = (req.headers.authorization || "").replace(/^Bearer /, "");
     Keyfetch.jwt
       .verify(jwt, verifyOpts)
       .then(async function (decoded) {
         if (decoded.claims.iss != verifyOpts.iss) {
+          // a failsafe for older versions of keyfetch with ['*'] by default
           throw new Error(
             `unexpectedly passed issuer validation: '${decoded.claims.iss}' does not match '${verifyOpts.iss}'`
           );
@@ -168,7 +100,9 @@ module.exports = function (issuer, PRIVATE_KEY, myOptions) {
   let exchangeClaims;
 
   let opts = {
+    issuer: issuer,
     oidc: {},
+    oauth2: {},
     notify: null,
     store: null,
     challengeMaxAge: 0,
@@ -177,8 +111,6 @@ module.exports = function (issuer, PRIVATE_KEY, myOptions) {
     DEVELOPMENT: "DEVELOPMENT" === process.env.ENV,
     _developmentSendChallengeSecret: false,
   };
-
-  let iss = issuer;
 
   // Cookie Stuff
   // See https://github.com/BeyondCodeBootcamp/beyondcodebootcamp.com/blob/main/articles/express-cookies-cheatsheet.md
@@ -255,7 +187,7 @@ module.exports = function (issuer, PRIVATE_KEY, myOptions) {
       claims = {};
     }
     if (!claims.iss && false !== claims.iss) {
-      claims.iss = iss;
+      claims.iss = opts.issuer;
     }
     return { claims, id_claims, access_claims, refresh_claims };
   }
@@ -272,7 +204,7 @@ module.exports = function (issuer, PRIVATE_KEY, myOptions) {
       );
       id_token = await Keypairs.signJwt({
         jwk: keypair.private,
-        iss: iss,
+        iss: opts.issuer,
         exp: id_claims.exp || defaultIdMaxAge,
         // optional claims
         claims: id_claims,
@@ -287,7 +219,7 @@ module.exports = function (issuer, PRIVATE_KEY, myOptions) {
       );
       access_token = await Keypairs.signJwt({
         jwk: keypair.private,
-        iss: iss,
+        iss: opts.issuer,
         exp: access_claims.exp || defaultAccessMaxAge,
         // optional claims
         claims: access_claims,
@@ -309,7 +241,7 @@ module.exports = function (issuer, PRIVATE_KEY, myOptions) {
     maxAge = refresh_claims.exp || defaultRefreshMaxAge;
     refresh_token = await Keypairs.signJwt({
       jwk: keypair.private,
-      iss: iss,
+      iss: opts.issuer,
       exp: maxAge,
       claims: refresh_claims,
     });
@@ -346,7 +278,7 @@ module.exports = function (issuer, PRIVATE_KEY, myOptions) {
     }
 
     return Keyfetch.jwt.verify(req.signedCookies.refresh_token, {
-      iss: iss,
+      iss: opts.issuer,
     });
   }
 
@@ -359,7 +291,7 @@ module.exports = function (issuer, PRIVATE_KEY, myOptions) {
     }
 
     let jws = await Keyfetch.jwt.verify(jwt, {
-      issuers: [iss], // or ['*']
+      issuers: [opts.issuer], // or ['*']
       jwk: pub,
     });
 
@@ -372,7 +304,7 @@ module.exports = function (issuer, PRIVATE_KEY, myOptions) {
       pub = keypair.public;
     }
     let jws = await Keyfetch.jwt.verify(jwt, {
-      issuers: [iss], // or ['*']
+      issuers: [opts.issuer], // or ['*']
       jwk: pub,
     });
     return jws;
@@ -411,7 +343,7 @@ module.exports = function (issuer, PRIVATE_KEY, myOptions) {
       );
       let jwt = await Keypairs.signJwt({
         jwk: keypair.private,
-        iss: iss,
+        iss: opts.issuer,
         exp: "24h",
         // optional claims
         claims: access_claims,
@@ -468,70 +400,12 @@ module.exports = function (issuer, PRIVATE_KEY, myOptions) {
         HMAC_SECRET: opts.secret,
         notify: opts.notify,
         store: opts.store,
-        iss,
+        iss: opts.issuer,
         _authnParam: opts.authnParam,
         _getClaims,
         _grantTokensAndCookie: grantTokensAndCookie,
       });
       app.use("/challenge", byChallenge);
-    }
-
-    if (process.env.GITHUB_CLIENT_SECRET) {
-      app.get("/webhooks/oauth2/github", async function (req, res) {
-        let request = require("@root/request");
-        // https://docs.github.com/en/developers/apps/building-oauth-apps/authorizing-oauth-apps#web-application-flow
-
-        let clientId = process.env.GITHUB_CLIENT_ID;
-        let clientSecret = process.env.GITHUB_CLIENT_SECRET;
-        let code = req.query.code;
-        // TODO check state
-        //let state = req.query.state;
-
-        let resp = await request({
-          method: "POST",
-          url: "https://github.com/login/oauth/access_token",
-          // www-urlencoded...
-          form: {
-            client_id: clientId,
-            client_secret: clientSecret,
-            code: code,
-            // TODO
-            //redirect_uri: process.env.GITHUB_REDIRECT_URI,
-          },
-        });
-        // TODO issuer may not be 1:1 with return url
-        //var url = new URL(`${issuer}?${resp.toJSON().body}`);
-        var url = new URL(`${issuer}#${resp.toJSON().body}`);
-        var token = url.searchParams.get("access_token");
-        url.searchParams.set("issuer", "github.com");
-        url.searchParams.set("state", req.query.state);
-
-        res.statusCode = 302;
-        res.setHeader("Location", url.toString());
-        res.end("<!-- Redirecting... -->");
-      });
-
-      let byOauth2 = async function (req, res) {
-        req[opts.authnParam] = {
-          strategy: "oauth2",
-          email: req._oauth2.email,
-          iss: req._oauth2.iss,
-          id: req._oauth2.id, // TODO
-          oauth2_profile: req._oauth2,
-        };
-        let allClaims = await _getClaims(req);
-        req[opts.authnParam] = null;
-        // TODO delete req._oauth2? (and for oidc too?)
-
-        let tokens = await grantTokensAndCookie(allClaims, req, res);
-        res.json(tokens);
-      };
-
-      app.post(
-        "/session/oauth2/github.com",
-        verifyGitHubToken(/*{}*/),
-        byOauth2
-      );
     }
 
     if (exchangeClaims) {
@@ -630,43 +504,34 @@ module.exports = function (issuer, PRIVATE_KEY, myOptions) {
       opts.logout = fn;
     },
     oauth2: function (_opts) {
-      // TODO
+      opts.oauth2 = Object.assign(opts.oauth2, _opts);
+      _init();
+
+      // TODO pick one!
+      var _gh = opts.oauth2["github.com"] || opts.oauth2.github;
+      if (_gh?.clientSecret) {
+        require("./oauth2/github.com/")({
+          app,
+          _gh,
+          opts,
+          _getClaims,
+          grantTokensAndCookie,
+        });
+      }
     },
     oidc: function (_opts) {
       opts.oidc = Object.assign(opts.oidc, _opts);
       _init();
 
-      let googleVerifierOpts = {};
-      if (opts.DEVELOPMENT && !opts.__DEVELOPMENT_2) {
-        opts.__DEVELOPMENT_2 = true;
-        if (opts.oidc?.google?.clientId) {
-          console.info(
-            "[auth3000] [ENV=DEVELOPMENT] Allow Expired Google Tokens"
-          );
-          // allow tests with expired google example token
-          googleVerifierOpts.exp = false;
-        }
-      }
-      if (opts.oidc?.google?.clientId) {
-        let byOidc = async function (req, res) {
-          req[opts.authnParam] = {
-            strategy: "oidc",
-            email: req._jws.claims.email,
-            iss: req._jws.claims.iss,
-            ppid: req._jws.claims.sub,
-            oidc_claims: req._jws.claims,
-          };
-          let allClaims = await _getClaims(req);
-          req[opts.authnParam] = null;
-
-          let tokens = await grantTokensAndCookie(allClaims, req, res);
-          res.json(tokens);
-        };
-        app.post(
-          "/session/oidc/google.com",
-          verifyGoogleToken(opts.oidc.google.clientId, googleVerifierOpts),
-          byOidc
-        );
+      var _goog = opts.oidc["accounts.google.com"] || opts.oidc.google;
+      if (_goog?.clientId) {
+        require("./oidc/accounts.google.com/")({
+          app,
+          opts,
+          verifyOidcToken,
+          _getClaims,
+          grantTokensAndCookie,
+        });
       }
     },
     options: function (_opts) {
@@ -683,8 +548,8 @@ module.exports = function (issuer, PRIVATE_KEY, myOptions) {
         "/.well-known/openid-configuration",
         async function (req, res) {
           res.json({
-            iss: iss,
-            jwks_uri: iss + "/.well-known/jwks.json",
+            iss: opts.issuer,
+            jwks_uri: opts.issuer + "/.well-known/jwks.json",
           });
         }
       );
